fix: stage before BuildSetup +semver: fix

+++ sidi.project-package.Test/test.ps1
@@ -2,10 +2,20 @@ $ErrorActionPreference = "Stop";
 
 $chocolateySource = Get-Item out\Release\Chocolatey | % { $_.FullName }
 
+function EnsureCdEmptyDirectory
+{
+	Param ([string] $dir)
+
+	if (Test-Path -Path $dir)
+	{
+		rm -r -fo $testDir
+	}
+	mkdir $dir
+	cd $dir
+}
+
 $testDir = "out\Release\test\end-to-end-test"
-rm -r -fo $testDir
-mkdir $testDir
-cd $testDir
+EnsureCdEmptyDirectory $testDir
 
 nuget install -OutputDirectory tool -Verbosity detailed -Source $chocolateySource sidi.project
 $toolDir = Get-Item tool\sidi.project.*
